Feat: :sparkles: Add mappers

+++ src/main/java/com/Campusland/ProyectoSpringBoot_CorpusEnrique/dto/response/DepartamentoResponse.java
@@ -1,7 +1,7 @@
 package com.Campusland.ProyectoSpringBoot_CorpusEnrique.dto.response;
 
 
-public record DepartamentoResponse(  Integer idDepartamento,
+public record DepartamentoResponse(  Long idDepartamento,
                                     String nombre) {
