+++ src/main/java/com/Campusland/ProyectoSpringBoot_CorpusEnrique/model/Departamento.java
@@ -11,7 +11,7 @@ public class Departamento {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_departamento")
-    private Integer idDepartamento;
+    private Long idDepartamento;
 
     @Column(name = "nombre", nullable = false, unique = true, length = 100)
     private String nombre;
